refactor: simplify version scripts by extracting common utilities
- Create shared version-utils.ts module with common functions
- Reduce calculate-next-version.ts from 207 to 125 lines (40% reduction)
- Reduce get-current-version.ts from 131 to 74 lines (43% reduction)
- Eliminate code duplication between scripts
- Improve maintainability and testability
- Update documentation to reflect new structure
- All functionality preserved and tested
